+++ custom_components/linknlink/const.py
@@ -4,5 +4,5 @@ from homeassistant.const import Platform
 
 DOMAIN = "linknlink"
 CONF_HOST = "host"
-PLATFORMS = (Platform.SENSOR, Platform.BINARY_SENSOR)
+PLATFORMS = (Platform.SENSOR, Platform.BINARY_SENSOR, Platform.EVENT)
 UPDATE_INTERVAL_SECONDS = 30